pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "django-deep-link"
description = "Simple app to get mobile app deep linking in place with basic tracking.."
classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
]
dependencies = ["ua-parser", "user-agents", "django-ipware", "requests"]

dynamic = ["version", "readme"]
requires-python = ">=3.8"
authors = [{ name = "Tim Santor", email = "tsantor@xstudios.com" }]

[project.optional-dependencies]
dev = ["black", "pre-commit"]

[tool.setuptools.packages.find]
# https://setuptools.pypa.io/en/latest/userguide/datafiles.html
where = ["src"]

[tool.setuptools.dynamic]
version = { attr = "django_deep_link.__version__" }
readme = { file = ["README.md", "HISTORY.md"], content-type = "text/markdown" }


# ==== pytest ====
[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--ds=django_project.settings --reuse-db"
python_files = ["tests.py", "test_*.py"]
norecursedirs = ["node_modules"]
filterwarnings = "ignore::DeprecationWarning"
pythonpath = "src"

# ==== Coverage ====
[tool.coverage.run]
include = ["*/django_deep_link/**/*.py"]
omit = ["*/migrations/*", "*/tests/*"]
plugins = ["django_coverage_plugin"]

# ==== mypy ====
[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
# Django migrations should not produce any errors:
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "django_project.settings"

# ==== djLint ====
[tool.djlint]
blank_line_after_tag = "load,extends"
close_void_tags = true
format_css = true
format_js = true
# TODO: remove T002 when fixed https://github.com/Riverside-Healthcare/djLint/issues/687
ignore = "H006,H030,H031,T002"
include = "H017,H035"
indent = 2
max_line_length = 119
profile = "django"

[tool.djlint.css]
indent_size = 2

[tool.djlint.js]
indent_size = 2

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
  "*/migrations/*.py",
  "staticfiles/*",
]
# Same as Django: https://github.com/cookiecutter/cookiecutter-django/issues/4792.
line-length = 88
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = [
  "F",
  "E",
  "W",
  "C90",
  "I",
  "N",
  "UP",
  "YTT",
  # "ANN", # flake8-annotations: we should support this in the future but 100+ errors atm
  "ASYNC",
  "S",
  "BLE",
  "FBT",
  "B",
  "A",
  "COM",
  "C4",
  "DTZ",
  "T10",
  "DJ",
  "EM",
  "EXE",
  "FA",
  'ISC',
  "ICN",
  "G",
  'INP',
  'PIE',
  "T20",
  'PYI',
  'PT',
  "Q",
  "RSE",
  "RET",
  "SLF",
  "SLOT",
  "SIM",
  "TID",
  "TCH",
  "INT",
  # "ARG", # Unused function argument
  "PTH",
  "ERA",
  "PD",
  "PGH",
  "PL",
  "TRY",
  "FLY",
  # "NPY",
  # "AIR",
  "PERF",
  # "FURB",
  # "LOG",
  "RUF",
]
ignore = [
  "S101",   # Use of assert detected https://docs.astral.sh/ruff/rules/assert/
  "RUF012", # Mutable class attributes should be annotated with `typing.ClassVar`
  "SIM102", # sometimes it's better to nest
  "ERA001",
  "PGH004",
  "E501",
]
# Project specifc ignore
extend-ignore = [
  "S106",   # hardcoded password in tests
  "RUF005", # iterable unpacking instead of concatenation
  "S104",   # Possible binding to all interfaces,
  "S311",   # We're not using it for cryptography
  "FBT001", # Boolean typed positional argument
  "FBT002", # Boolean default positional argument
  "S308",   # mark_safe is ok here
  # "SLF001"  # accessing private _meta is ok here
  "TRY300", # Consider moving this statement to an `else` block
  "COM812", # Recommended from ruff
  "ISC001", # Recommended from ruff
]
# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.isort]
force-single-line = true
